+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 This project uses [semantic versioning](https://semver.org/spec/v2.0.0.html)
 
 
+## [0.1.1] 2021-03-29
+### Added
+- Create directory doc
+- Add file usage in directory doc
+- Create package qr
+- Create files decode.go and encode.go
+- Create directory test
+- Adition files tests
+### Changed
+- Moved file version for directory doc
+- Moved logic the program for package qr
+
 ## [0.1.0] 2021-03-26
 ### Added
 - Adition of README.md
@@ -14,3 +26,4 @@ This project uses [semantic versioning](https://semver.org/spec/v2.0.0.html)
 - Created initial cli
 
 [0.1.0]: https://github.com/informeai/genqrcode/releases/tag/v0.1.0
+[0.1.1]: https://github.com/informeai/genqrcode/releases/tag/v0.1.1